Move preferences-dialog-related operations from Termonad.App to Termonad.Preferences
Termonad.Preferences also re-exports helpful funtions from
Termonad.Preferences.File.
